# client/public/swagger/swagger.yaml
openapi: 3.0.3
info:
  title: BRAAAAAM API
  version: 1.0.0
  description: BRAAAAAM API
servers:
  - url: https://api.braaaaam.webdev.gccis.rit.edu
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    OrgId:
      name: orgId
      in: path
      required: true
      schema: { type: string }
...
  schemas:
    Error:
      type: object
      properties:
        message: { type: string }
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
paths:
...
  /o/{orgId}/users:
    get:
      tags: [Users]
      summary: List users
      parameters: [ { $ref: '#/components/parameters/OrgId' } ]
      security: [ { bearerAuth: [] } ]
      responses:
        '200': { description: OK }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Users]
      summary: Create user
      parameters: [ { $ref: '#/components/parameters/OrgId' } ]
      security: [ { bearerAuth: [] } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [firstName, lastName, email, phone]
              properties:
                firstName: { type: string }
                lastName:  { type: string }
                email:     { type: string }
                phone:     { type: string }
      responses:
        '201': { description: Created }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /o/{orgId}/clients:
    get:
      tags: [Clients]
      summary: List clients
      parameters: [ { $ref: '#/components/parameters/OrgId' } ]
      responses:
        '200': { description: OK }
    post:
      tags: [Clients]
      summary: Create client
      parameters: [ { $ref: '#/components/parameters/OrgId' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [firstName, lastName, contactPreference, vehiclePreferenceType, notes, gender]
              properties:
                firstName: { type: string }
                lastName:  { type: string }
                email:     { type: string }
                phone:     { type: string }
                contactPreference:     { type: string }
                addresses:
                  type: array
                  items:
                    type: object
                    required: [addressLine1, city, state, zip, country, isPrimary]
                    properties:
                      addressLine1: { type: string }
                      city:         { type: string }
                      state:        { type: string }
                      zip:          { type: string }
                      country:      { type: string }
                      isPrimary:    { type: boolean }
                      vehiclePreferenceType: { type: string }
                      notes: { type: string }
                      gender: { type: string, enum: [Male, Female] }
      responses:
        '201': { description: Created }
...
  /o/{orgId}/settings:
    get:
      tags: [Settings]
      summary: Get settings
      parameters: [ { $ref: '#/components/parameters/OrgId' } ]
      responses:
        '200': { description: OK }
    put:
      tags: [Settings]
      summary: Update settings
      parameters: [ { $ref: '#/components/parameters/OrgId' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                hoursOfOperation:
                  type: array
                  items:
                    type: object
                    required: [Id, dayOfWeek, startTime, endTime]
                    properties:
                      id:        { type: string }
                      dayOfWeek: { type: integer, description: "0-6" }
                      startTime: { type: string }
                      endTime:   { type: string }
                name:               { type: string }
                logo:               { type: string }
                domain:             { type: string }
                phone:              { type: string }
                email:              { type: string }
                streetAddress1:   { type: string }
                streetAddress2:   { type: string }
                city:               { type: string }
                state:              { type: string }
                zip:                { type: string }
                country:            { type: string }
                postmarkApiKey:     { type: string }
                googleMapsApiKey:   { type: string }
      responses:
        '200': { description: Updated }
...
  /o/{orgId}/settings/roles:
    get:
      tags: [Roles]
      summary: List roles
      parameters: [ { $ref: '#/components/parameters/OrgId' } ]
      responses:
        '200': { description: OK }
    post:
      tags: [Roles]
      summary: Create role
      parameters: [ { $ref: '#/components/parameters/OrgId' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [roleName, description, permissionIds]
              properties:
                roleName:     { type: string }
                description:  { type: string }
                permissionIds:
                  type: array
                  items: { type: string }
      responses:
        '201': { description: Created }
...
  /o/{orgId}/settings/locations:
    post:
      tags: [Locations]
      summary: Create location
      parameters: [ { $ref: '#/components/parameters/OrgId' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [address, alias]
              properties:
                address:
                  type: object
                  required: [addressLine1, city, state, zip, country]
                  properties:
                    addressLine1: { type: string }
                    addressLine2: { type: string }
                    city:           { type: string }
                    state:          { type: string }
                    zip:            { type: string }
                    country:        { type: string }
                Alias: { type: string }
      responses:
        '201': { description: Created }
...
  /o/{orgId}/appointments:
    get:
      tags: [Appointments]
      summary: List appointments
      parameters: [ { $ref: '#/components/parameters/OrgId' } ]
      responses:
        '200': { description: OK }
    post:
      tags: [Appointments]
      summary: Create appointment
      parameters: [ { $ref: '#/components/parameters/OrgId' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [startDate, startTime, estimatedEndDate, estimatedEndTime, client, pickupLocation, dropoffLocation, status]
              properties:
                startDate:         { type: string }
                startTime:         { type: string }
                estimatedEndDate:  { type: string }
                estimatedEndTime:  { type: string }
                client:
                  type: array
                  items: { type: string, description: "@client id(s)" }
                pickupLocation:     { type: string, description: "@location id" }
                dropoffLocation:   { type: string, description: "@location id" }
                status:            { type: string }
      responses:
        '201': { description: Created }
...
  /o/{orgId}/notifications:
    get:
      tags: [Notifications]
      summary: List notifications
      parameters: [ { $ref: '#/components/parameters/OrgId' } ]
      responses:
        '200': { description: OK }
...
  /o/{orgId}/reports:
    get:
      tags: [Reports]
      summary: List reports
      parameters: [ { $ref: '#/components/parameters/OrgId' } ]
      responses:
        '200': { description: OK }
    post:
      tags: [Reports]
      summary: Create report
      parameters: [ { $ref: '#/components/parameters/OrgId' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, query]
              properties:
                title: { type: string }
                query: { type: string }
      responses:
        '201': { description: Created }
...